move Sidepanel to Editor feat
